[automatic failover] Move failover provider to mcf (#4294)
* - move failover provider to mcf

* - make iterateActiveCluster package private
