+++ src/test/java/redis/clients/jedis/misc/AutomaticFailoverTest.java
@@ -18,8 +18,9 @@ import redis.clients.jedis.*;
 import redis.clients.jedis.exceptions.JedisAccessControlException;
 import redis.clients.jedis.exceptions.JedisConnectionException;
 import redis.clients.jedis.mcf.ClusterSwitchEventArgs;
+import redis.clients.jedis.mcf.MultiClusterPooledConnectionProvider;
+import redis.clients.jedis.mcf.MultiClusterPooledConnectionProviderHelper;
 import redis.clients.jedis.mcf.SwitchReason;
-import redis.clients.jedis.providers.MultiClusterPooledConnectionProvider;
 import redis.clients.jedis.util.IOUtils;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -75,7 +76,8 @@ public class AutomaticFailoverTest {
       AbstractPipeline pipe = client.pipelined();
       pipe.set("pstr", "foobar");
       pipe.hset("phash", "foo", "bar");
-      provider.iterateActiveCluster(SwitchReason.HEALTH_CHECK);
+      MultiClusterPooledConnectionProviderHelper.iterateActiveCluster(provider,
+        SwitchReason.HEALTH_CHECK);
       pipe.sync();
     }
 
@@ -94,7 +96,8 @@ public class AutomaticFailoverTest {
       AbstractTransaction tx = client.multi();
       tx.set("tstr", "foobar");
       tx.hset("thash", "foo", "bar");
-      provider.iterateActiveCluster(SwitchReason.HEALTH_CHECK);
+      MultiClusterPooledConnectionProviderHelper.iterateActiveCluster(provider,
+        SwitchReason.HEALTH_CHECK);
       assertEquals(Arrays.asList("OK", 1L), tx.exec());
     }
 
@@ -151,8 +154,8 @@ public class AutomaticFailoverTest {
     MultiClusterClientConfig.Builder builder = new MultiClusterClientConfig.Builder(
         getClusterConfigs(clientConfig, hostPortWithFailure, workingEndpoint.getHostAndPort()))
             .retryMaxAttempts(retryMaxAttempts) // Default
-                                 // is
-                                 // 3
+            // is
+            // 3
             .circuitBreakerSlidingWindowMinCalls(slidingWindowMinCalls)
             .circuitBreakerSlidingWindowSize(slidingWindowSize);
